Default to nic cld_rst override asserted after reset
Gimlet B requires SP to notice the SP3 enabling power
to the nic. A future improvement opportunity would be
to wire this directly to the FPGA which is controlling
all of the NIC sequencing anyway.

As such, right now CLD reset de-asserts when nic is
powered and latches DBG0 state, which is *way* before
SP3 has a chance to actually set it, so for now we
default our dbg override bit to asserted and manually
come in and clear it when we want to later. A better
API for this should be developed and integrated but
it's bring-up and this was the fastest way to get it
done now.