Implement Sidecar PCIe endpoint control, refactor PCIe software reset
This diff adds basic software control for the PCIe endpoint control signals
terminating on the board controller. This allows the software sequencer control
over the present, reset, fault and alert signals between Gimlet and Sidecar.

In order to stick to the "one point of control" philosophy software control over
the PCIe reset signal was removed from the Tofino 2 sequencer. The software
sequencer is instead expected to use the `OVERRIDE_HOST_RESET` and `RESET` bits
in the endpoint `CTRL` register. As a bonus this slightly reduces complexity in
`Tofino2Sequencer`.

Test cases have been reworked accordingly.