Refactor Sidecar mainboard controller top, add mainboard emulator (#12)
This is a refactor of the Sidecar mainboard controller top, which does the
following:

Flatten the interfaces to a single Top interface and module which deals with
syncing discrete signals, connecting them to a slightly more type rich
MainboardController module, which then further breaks functionality into smaller
pieces. The Top module isn't pretty, but it keeps all the gory details of I/O
in a single place, allowing for a clearner, discrete MainboardController module
which is usable in test benches etc.

In addition a basic emulator targeting the ECP5 EVN board is provided, which
emulates the Tofino PDN and allows the sequencer to run via a connected
Gimletlet. This may very well migrate to the partial QSFP board in the future
since the EVN boards are a bit hard to get.

Finally, this diff has an example of a more minimal SPI interface, making use
of the register interface of various submodules implemented using ConfigRegs.