Refactor Sidecar mainboard controller register map, add version/checksum registers
This diff refactors the MainboardControllerSpiServer to accommodate for a larger
register map by reducing the fan-out/fan-in of requests/responses during a
single cycle. This was forward ported from upcoming Ignition diffs.

In addition the current implementation allows for ActionValue registers, needed
to implement for example counter registers which reset upon being read.

As part of the refactor the Git version and checksum registers are added to the
first page, to aid the sequencer in (re)loading the bitstream on boot.