Improve VSC8562 fault handling, add simulations (#74)
This commit adds actual fault handling for if the VSC8562 power supplies cannot be turned on successfully, closing #60. The previous code would simply wait forever, but now a timeout will abort the power sequencing until the timeout has been cleared via the PHY_CTRL.CLEAR_POWER_FAULT bit. I've also added a suite of simulations to verify the power sequencing functionality as well as the SMI (MDIO) register interface. While all these changes are a superset to the original code, I took the opportunity to standardize this module's multi-byte registers to be little-endian, matching the rest of the design. This is a breaking change to Hubris that should be addressed when this design is pulled in. Finally, the power good inputs for both of the PHY power supplies have been changed to open-drain inputs with internal pullups to reflect the actual IO structure.

It also fixes a bug in the MDIOPeripheralModel where the actual data it was transmitting was bit 0 to bit 15, rather than bit 15 to bit 0. This was not caught because in the simulation I had an oversight where I was only checking what the model said it was transmitting, not what it it was actually transmitting. This has also been fixed in MDIOTests.