FMC to AXI interface (#138)
This adds an FMC to AXI interface and sim environment. The sim env
provides an FMC model based on our understanding of the FMC interface
and a programming interface to interact with the FMC model to issue
read/write transactions.

The sim model is then wired up to the FMC target which does the FMC to
AXI-Lite transformation. This unit currently supports 32bit
transactions, but is designed such that adding FMC-side bursting
capability should be fairly straight forward once we have a use-case,
but that work has been left for a future development. The target does
support clock crossing from the FMC domain into the AXI clock domain,
and has FIFOs present for synthesizing longer transactions in a more
efficient manner. The documentation describes some potential challenges
around implementing bursting, notably that it isn't possible to do
proper AXI-compliant bursting in the write-direction without store and
forwarding the writes, or doing some kind of chunking onto smaller AXI
bursts. On the read side, it's a similar story in that we don't know the
burst size up front, so no good bursting strategy can be done.

Closes #143

---------

Co-authored-by: Aaron Hartwig <aaron@oxidecomputer.com>