Bump oss-cad-suite to latest (#150)
Our current toolchain is pinned to a release of the OSS CAD Suite that
is ~1.5 years old. A lot of progress has been made in that time, and
using the QSFP design as a spot check, we can get quite the benefit from
an update. On that design I saw a ~32% reduction in utilization and a
45% increase in Fmax, all while actually completing PnR 7x faster.

I'm opening this PR to generate images in CI. I'll look at all the PnR
reports and make sure they make sense, then I'll pull images into Hubris
branches and let them soak on `niles` and see if anything quirky
happens. I'll be testing:

* Sidecar Mainboard Controller
* QSFPx32 Controller
* Gimlet Sequencer
* Ignition Target (I'll ask Eric to flash this on any sleds he can)

If we aren't seeing any issues after a few days, I'll explore
integrating these into Hubris and try to get the new images released in
the dogfood/colo context ASAP to get a good long soak before R9.