Allow 8, 16, or 32 bit registers in RDL and generated files (#156)
Previously the tool was limited (only by pretty html generation) to 8
bit register widths. We'd like to support 32bits for cosmo and adding 16
was easy to round out the supported widths.

Mixing of widths in the same rdl generation is out of scope and wasn't
tested as we don't anticipate that being a case we'll need right now.