Figure out generated vhdl files from rdl and add to lsp config (#160)
Previously, the generated vhdl package files from the rdl spec were
missing in the files passed to the VHDL-LS. These are the only
non-source files we're dealing with so we need the bxl action to
manifest these locally and dealing with generated files is just a bit
different than input sources from BXL's perspective.

This change generates the outputs as part of the BXL action, ensures
they're local and then feeds that into the data structure that feeds the
lsp.

It also provides a new subcommand to the multitool which wraps this up
and obviates the need for a standalone json->toml converter, but does
introduce a new 3rd party python dependency (tomli-w). The tool errors
out with a nice suggestion if you attempt to run it without the proper
python package installed.