Support specification of VHDL standards for sim and synth (#167)
Our simulation tools can't support multiple standards in the same
simulation, this means that for simulation, we have to "lift" all of the
files up to the maximum used VHDL standard. This PR addresses this.

Vivado can deal with mixed standards, but we didn't support that either,
defaulting to 208 so this PR addresses this as well.

In order to make all of this work, additional properties had to be
plumbed through, and this allows our synthesis generator to detect and
drop non-synth files, which is something we needed to do long-term
anyway. The big use-case here is that for simulation we're using VHDL
models of the XPM library, but Vivado will provide those natively. We'd
have dependencies in the tree for these models which we'd need to strip
when making the list of files to synthesize. This PR resolves this as
well since the features now exist to to the other plumbing.

Finally, we default the vhdl-lsp to 2019 mode, which should be the most
useful editor mode. This does mean the LSP could detect valid 2019
syntax that a synth or sim tool doesn't yet support but there's not a
lot to do about that.