Sidecar: enable PU on the PERST input pin, add rev D in build target name (#201)
As per https://github.com/oxidecomputer/hardware-sidecar/issues/830 we
want to enable the internal pull-up on the mainboard controller for the
pin connected to the buffered PERST signal. This diff does so for all
revs (B, C, D). In addition the build target now reflects rev D
compatibility which was never committed 7 months ago :(