UART and FIFO UART blocks (#206)
A simple AXI streaming uart and fifo'd uart block with a functional
testbench. More cases can be added but the initial test sends a byte
into the serial interface and has the streaming interfaces looped back
so that it sends the byte back out the serial interface, and goes
through the FIFOs etc.