Make register records use std_logic_vector types vs unsigned
This always felt a little gross but by and large users wanted to
use multi-bit fields as counters etc so making them unsigned 
was an easy way to facilitate that. With the advent of 
ieee.numeric_std_unsigned support, consumers of std_logic_vectors
can now opt-in to "unsigned" behavior of the std_logic_vectors
when desired, and this works on tools that support VHDL2008+ so
we make records slv types again. This represents an interface change
but this change includes the only block using this and already in
mainline quartz right now.