I2C: add support for clock stretching (#210)
This adds support for clock stretching to the `I2CBitController`. This
block now also exposes `scl_stretch_seen` and `scl_stretch_timeout`
signals to the `I2CCore` which in turn exposes them to application
logic. Additionally, if a stretching timeout occurs, the bit controller
and the core will clear their state and wait for a new transaction to
occur. The `I2CCore` will also perform acknowledge-polling on all write
operations performed before considering the operation complete.

The `QsfpModuleController` will pay attention to the timeout signal and
yield a new `PortStatus::Error::I2cSclStretchTimeout` should it occur.
It will also expose `scl_stretch_seen` via `PortStatus::StretchingSeen`.

Finally, there was some touching in the Sidecar mainboard controller
design to add in the required parameters to the `TofinoDebugPort` block
which is a user of the `I2CCore`. These changes should be transparent as
does not clock stretch on that interface.

Fixes #114 
Fixes #115