Utilities: Adds a Strobe block (#224)
This PR adds a simple strobe block that issues a single-cycle pulse
every `TICKS` clk periods. The intention behind it is you could
calculate a constant for `TICKS` using the `time_pkg` to get pulses at
various frequencies. I reorganized the utilities folder a bit so we
don't end up with a bunch of files just laying around in that directory.
I'm open for discussion on organization here.

The PR also creates a `common` directory which is intended to house
reusable blocks and packages. I added a `streaming_if_pkg` that has
a generic width streaming interface definition.

Finally, there are a few small fixes to the tb/th generation templates.