Espi boot integration on ruby + grapefruit (#230)
This contains a number of bugfixes and changes identified while working
through ruby + grapefruit espi boot, culminating in a successful ruby
espi boot!

This is kind of a mess as there were a number of issues identified
throughout the associated blocks.
A few notes on major changes/bugfixes:
- Sync async inputs
- Use grapefruit espi_alert_l pin as DAT[1] per AMD's guide.
- eSPI serializer defaults to '1' at the last sclk fedge of a
transaction. This is per eSPI spec and helps return the bus high given
the weak pullups so we don't get spurious requests.
- In-band alert logic was updated to latch live status before sending
the status bits, and to abort a pending alert if a current status word
was shifted out.
- Fixed 0-index/1-index discrepancy when generating spi nor
transactions. SPI block uses 1-index for a more clear hubris api, so we
need to adjust things appropriately when generating and executing flash
transactions from eSPI
- Added VWire channel support. The SP5 mask rom unconditionally does a
set-vwire to indicate platform reset and we need to handle it.
- Stop advertising OOB support, we don't support this and the ASP image
when autoconfiguring was (rightly) upset that we advertised support but
wouldn't allow it to be enabled
- Respond to PUT_FLASH_NP with DEFER instead of ACCEPT: ACCEPT means you
have the data immediately available which is not the case when we have
to go fetch it from flash.
- Fix bug where we'd read an empty fifo from the spi-nor block when
filling a DPR queue as part of the flash channel response logic.
- Some simulation fixes and improvements.