I2C: A couple of fixes around SCL stretching is handled (#248)
This commit fixes a couple of bugs related to SCL stretching. First off,
in the `I2CBitController` I adjusted how we sample SCL to detect
stretching so we're not errantly sampling. Then in the `I2CCore` I added
a mechanism to handle the SCL stretch timeouts when we see them. Before
we lacked this and that meant if that was seen, we were wedged. The rest
of the changes were to test cases in order to mimic the behavior
observed that brought the problems to our attention.

Fixes #246
Fixes #247