I2C controller: add ability to abort transactions (#279)
A feature we need for the SPD mux testing is the ability for our
controller to abort an in-progress transaction. This commit adds an
`abort` port to the `i2c_ctrl_txn_layer` entity which will react to a
pulse of that signal. In practice this means if the controller is
writing it will send a STOP instead of the next bit. When reading, the
controller will finish reading the byte, NACK, and then send a STOP.