QSFP Controller: make power control and readback separate (#282)
The current state of upstack software understanding module power control
is confusing. The actual hardware power control had an additional level
of gating the FPGA that enforced a module being present before turning
power on. However, when software comes by to read the `POWER_ENx`
registers it sees the modules as being enabled. This is confusing!

This PR splits out the software control of the power enable into a
separate `SW_POWER_ENx` registers from the true pin-level hardware state
(post-gating) shown in the `POWER_ENx` registers.

The software impact of this changes are minimal. When fetching state the
`POWER_ENx` register will represent the what a user would expect. When
modifying if a power _should_ be enabled or not, the `SW_POWER_ENx`
registers will need to be hit instead.

fixes #280