Fix up some build flow issues.
1) If a design had multiple blocks that depend on the same core component
that also happened to have an RDL provider, when trying to generate the
maps at the end of the build we'd get failures b/c we just loop on the
RDLproviders and expected an output for each. That's fine, but when
there were multiple of the same provider, we'd try to re-declare outputs
and get errors from buck2. We take the list of files and turn it into a
set() which drops any duplicates before we loop them to generate our
outputs. We only want one copy of each anyway, so this is perfect.
2) Report stderr (ie ghdl warnings) *before* bailing on the return code so we can see the errors when it fails.
3) make the yosys_synth.py a little more maintainable.