Add an SPD proxy mux (#294)
This commit adds the `spd_proxy_top` block whose intention is to house
an I2C controller which will attempt to utilize the DIMM bus while a CPU
is not using it. When the CPU begins using the bus the internal
controller will wrap up what it is doing as quickly as possible and we
will attempt to gracefully hand the bus over the CPU in a way that
doesn't cause problems.