RDL naming convention enforcement and updates (#302)
Working with @mkeeter we're adding some improvements to the RDL
generation and I've made some naming convention enforcement to help
facilitate programatic consumption of the generated artifacts.

New Features:
- added "orig_type_name" to top level addrmap json so that the instance
name can be linked with the actual hardware generated .json. Remember,
we may have multiple copies of the same "hw" with different instance
names.
- added "addr_span_bytes" to the top level addrmap json to give a hint
to sw tool generation how many registers we have specified. Note that
there's a missing feature from SystemRDL spec that would allow reserving
undefined register space, so we're stuck telling sizes that are based
only on what we've defined. From a hw implementation perspective, it's
certainly safe to assume the peripheral owns up to the next power of 2
here, and possibly with a floor of 256byte addresses.

New Enforcement:
- To provide some naming consistency, we're now enforcing that buck2 rdl
targets end with "_rdl" so that it's easier to reason about, and then
enforce that output file names (which are user-specified) conform to the
expected patterns for software consumption. There are more improvements
to be made here https://github.com/oxidecomputer/quartz/issues/301 but
this is a functional start.

The new enforcement regime required doing some updates to various things
as seen in this PR as well.