make I2C controller push-pull (#307)
As discussed, the controller logic will now operate as a push-pull
driver. This gives flexibility since if that is desired it will be the
default, and adapting a push-pull driver to be open-drain is simple when
accompanying tristate logic is still present.

This change also changes the target VC to utilize the `*_int` signals
for rising/falling edge triggers since those are adapted to '0' or '1'
from the std_logic inout pins.