grapefruit: add spd proxy passthrough (#313)
This plops down the SPD block in grapefruit with the commands to the
internal FPGA controller disabled, effectively making it a CPU<->DIMM
passthrough (exercising the SDA arbitration logic at low speeds, at
least). The ruby boots with this image in place (all the activity shows
the reworked G channel 0 ACKing and all the ensuing traffic):

![image](https://github.com/user-attachments/assets/622eac46-1b8c-4568-83dc-f43704f0cb7c)

I get to nanoblrs successfully.

My plan is to merge this and get it put into @citrus-it's Hubris dev
branch. This means no more shuffling DIMMs around for you
@nathanaelhuffman!