SPD proxy work (#346)
This is a partially finished feature, but given the impact of all of the
changes to the VC's here I figured I'd get this mainlined and then
continue on with additional cosmo-specific feature work.

I also added a echo line before running the sims so it's easier to
figure out which sim command was running when sims fail.

This is a fairly extensive re-write of the i2c target vc, but all the
sims are running again without being adjusted and things are working
better than they did originally.