Add 8b10b encoders/decoders  (#351)
based on chuck benz verilog example.

Added some helper functions which are expected to be useful for ignition
simulations where this will be properly exercised, but this unit felt
stand-alone enough to be it's own foundational PR.

Rather than build out a bunch of simulations for these blocks (which are
effectively just math functions) I'd rather focus on integrating this
stuff into a more useful ignition simulation and use those test vehicles
to be the ongoing tests of these modules.

---------

Co-authored-by: Aaron Hartwig <aaron@oxidecomputer.com>