Make SP-console the default UART (#358)
Previously, we defaulted to shoving the SP5 uart data out the debug
header. That presented challenges with the mfg stations where a uart
dongle isn't necessarily connected leading to a wedge due to
hardware-handshake. This makes the default the SP which will keep the
buffer clear (like it does on gimlet) and we have a new memory mapped
peripheral to flip the mux back to the debug headers if we need it.

This lays the groundwork for additional debug control of various things
which we may want eventually also.

I booted this on my cosmo, and once booted I could use the `humility
console-proxy attach` command to get proper uart coms. This verified
that we don't wedge during boot and the the sp-console-proxy is working
as expected.