cosmo ignition target (#360)
I'm opening this up to start reviewing since this is a pile of code.

I'm  working to get this tested on hardware. I think this has
There's a lot of opportunity for additional testcases here, especially
on the target -> controller message direction as I haven't fully
implemented the controller rx path yet as I'm trying to determine the
proper api here but wanted to get some in-hardware testing going
concurrently given the timeline for needing *something* to work in the
lab here.

Closes #328