Some cosmo fix roll-ups (#371)
Fixes #364 . Baud rate generator was naive and would work fine for much
slower baud rates but as we get up into the MHz, it didn't provide
enough granularity. This is adjusted in the following ways:
- New generic clks_per_bit so we get precise bit timing
- Split tx and rx paths to use their own strobes again to provide more
precision in sampling and transmitting.

I also adjusted things so that the UARTs don't sample packets when we're
not in A0 to prevent accidentally filling up the fifos with noise or
non-actionable things.

This also includes a minor fix that was noticed during ignition flash
work where the front fpga doesn't release it's SPI output when it isn't
selected. This isn't strictly a problem with our design but I fixed it
to be a good spi citizen anyway.