Make the spd-ctrl bit hold until the hw prefetch finishes (#372)
Need to match the description of the spd_start bits so that they stay
set until the pre-fetcher completes.

Also de-buss the sp5_debug pins so we can more easily control the
direction. We'll be using sp5_to_fpga1_debug2 as an input for a bit to
deal with debugging t6 sequencing so I adjusted that here, it was
previously a tris-output so there's not flag-day on directionality.