add register to allow hubris to control whether we emit backplane PCIE clock when something is connected. (#378)
We're going to let lab images enable this in an upcoming PR so minibar
nics will work but we don't transmit the pcie clock out the back any
time a sidecar is connected.

This also includes previously un-merged UART debug that doesn't really
affect anything functionally but is there for getting some more info out
of the box for the host os team.