Better modularize the debug header (#379)
This sets up an arbitrary condiment dispenser for the 1v8 debug header.
I've grouped the pins in sets of 2 and you can mux out any of a few
interfaces to these pins by setting the correct bits in the mux
registers. ESPI is special in that it has a defined pinout, and there
are convenience registers to set all of this up correctly in one shot.

The false paths to the debug pins were already in place.
We have to clock this in the 200MHz domain
(since we want to faithfully represent espi), so we have false-path
entries to the debug pins to make timing work. We can see if this is a
problem long-term but I think it's not.

This also provides a means to grow our allowable, and run-time
selectable debug and allows concurrent debug of multiple interfaces,
limited by having to have pairs of signals and the number of pins
available.