Update to current buck2 2025-07-01 tag and prelude (#383)
This does required updating of the buck2 binary as described in the
README. This PR includes the prelude submodule update matching the tag.

Note that this didn't *specifically* break anything that I've noticed
but you should update your buck2 bin accordingly.

I made a no-change delta in a common file to force a rebuild of
everything since just adjusting the readme and the submodule don't
trigger our builds.

I specifically want a newer set of buck2 content to begin evaluation of
remote caching, which was unavailable until very recently:
https://github.com/facebook/buck2/pull/763 and friends.