wait until i2c stop to update mux (#386)
previously i2c mux actions were taken immediately on the write. In a
properly working system, this is a don't care usually but probably not
correct. And definitely in a system where the segment we're turning on
is already unhappy this breaks the saleae decode and makes debugging
harder.

All of the silicon i2c muxes I can find, including the 9545 state
something along the lines of:

> When a channel is selected, the channel will become active after a
STOP condition has
> been placed on the I2C-bus. This ensures that all SCx/SDx lines are in
a HIGH state
> when the channel is made active, so that no false conditions are
generated at the time of
> connection.

So we fix that here, and adjust the simulations to wait for the
transaction to completely finish before enabling the selected channel.

Fixes #387