consolidate our i2c muxes into one big 15 channel mux. (#392)
I'm working with matt on this implementation but it adjusts our mux
logic to support a 2byte register and deals with the exclusivity there.

This design will NACK invalid writes at the time of the 2nd byte is
written and we can perform exclusivity validation here.

This PR brings this IP into the repo, and I'll do a 2nd PR implementing
this in cosmo.

In the process I found a few issues with both our simulation controller
and the target logic to deal with multi-byte writes/reads, and making
sure that on reads the controller NACKs the final byte so the target
knows to be done.