minibar: make power button PC-like (#423)
The intention behind the power button is for it to provide a normal
PC-like behavior. When
power is disabled, a short press can turn power on. When power is
enabled, a user must intentionally press and hold the button for a
longer period (2 seconds) to get the system to disable power.

fixes #422