put ignition rev2+ builds into play (#426)
This puts the changes in place for rev2+ cosmo ignition images. This
also creates a new fpga build for rev1 which is essentially orphaned due
to different pin config.

In building this, we had a timing failure in oximux 16 on the cosmo-hp
build so I adjusted the 1hot detection logic to be more logic-efficient.