espi post code ring buffer (#432)
This ended up being a bit of an adventure. The implementation of the
buffer was actually very straight forward, but we had not to this point,
supported mixed size peripherals so we had to deal with that as well.

The solution represented here is functional but not optimal: This makes
the "responders" interface from the fabric all equivalently sized to 32
bit and the implementer needs to "trim" the axil interface down at point
of use. This allows the fabric to use loops over the same-size datatypes
and we can make overloaded procedures to trim them down to the
appropriate size for a given peripheral. This does foist some work on
the implementer so it's not quite as clean and plug-and-play as we'd
like- such is the cost of complexity.

The
[axilite_if_2k19_helper_pkg.vhd](https://github.com/oxidecomputer/quartz/compare/ndh/cosmo-postcode-buffer?expand=1#diff-bab8a559edf732b63a346c8e58be537f421eb396bced9a6842f501f3f918c42d)
contains examples of the overloading of trimming providers. I couldn't
quite get generic procedures + views working otherwise this could have
been a single generic procedure, as it is, you have to copy-paste it for
each target size you want to support :/

Then we implement the post-code stuff: we keep the "last seen" register
but add a counter and the buffer. Both counter and buffer pointers are
reset on an espi-reset, so they remain useful even after an A0->A2
transition, only resetting at the beginning of the AMD espi boot
sequence when the espi-reset is issued. (The buffer is ram, so its
contents aren't zeroed).

With some hubris patches to support the memory array, this is working on
my cosmo A:

The counter counts during boot:
```
➜  hubris git:(d5a8d559) ✗ humility hiffy -c FmcDemo.peek32 -a addr=0xc0008024
humility: attached via CMSIS-DAP
FmcDemo.peek32() => 0x491
➜  hubris git:(d5a8d559) ✗ humility hiffy -c FmcDemo.peek32 -a addr=0xc0008024
humility: attached via CMSIS-DAP
FmcDemo.peek32() => 0x4ac
```

last-seen is stable as before (we're booted at this point so no more
post-codes):
```
➜  hubris git:(d5a8d559) ✗ humility hiffy -c FmcDemo.peek32 -a addr=0xc0008020
humility: attached via CMSIS-DAP
FmcDemo.peek32() => 0xee110005
➜  hubris git:(d5a8d559) ✗ humility hiffy -c FmcDemo.peek32 -a addr=0xc0008020
humility: attached via CMSIS-DAP
FmcDemo.peek32() => 0xee110005
```

spot-checking buffer entries shows presumably valid payloads:
```
➜  hubris git:(d5a8d559) ✗ humility hiffy -c FmcDemo.peek32 -a addr=0xc0008100
humility: attached via CMSIS-DAP
FmcDemo.peek32() => 0xee1000b2
➜  hubris git:(d5a8d559) ✗ humility hiffy -c FmcDemo.peek32 -a addr=0xc0008104
humility: attached via CMSIS-DAP
FmcDemo.peek32() => 0xee1000a0
➜  hubris git:(d5a8d559) ✗ humility hiffy -c FmcDemo.peek32 -a addr=0xc0008108
humility: attached via CMSIS-DAP
FmcDemo.peek32() => 0xee1000a3
```

And a full boot sequence lands at about 0x51f codes:
```
➜  hubris git:(d5a8d559) ✗ humility hiffy -c FmcDemo.peek32 -a addr=0xc0008024
humility: attached via CMSIS-DAP
FmcDemo.peek32() => 0x51f
```

I took the unit to A2, verified none of this stuff changed and the
booted to A0 again and by the time I could look at it, we had some
post-codes, but a lower number than 0x51f showing that the counter
resets properly.

Fixes #419.